refactor(ChessBoardGenerator): ChessBoardGenerator 패키지 변경
